moved test helper functions per Jonathan Hehir's suggestion
